Factor out safe-mix and dispatch
