+++ demo/runtime/Cargo.toml
@@ -9,6 +9,7 @@ hex-literal = "0.1.0"
 log = { version = "0.3", optional = true }
 serde = { version = "1.0", default_features = false }
 serde_derive = { version = "1.0", optional = true }
+safe-mix = { path = "../../safe-mix", default_features = false}
 substrate-codec = { path = "../../substrate/codec" }
 substrate-runtime-std = { path = "../../substrate/runtime-std" }
 substrate-runtime-io = { path = "../../substrate/runtime-io" }
@@ -29,5 +30,6 @@ std = [
 	"demo-primitives/std",
 	"serde_derive",
 	"serde/std",
-	"log"
+	"log",
+	"safe-mix/std"
 ]
