+++ safe-mix/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "safe-mix"
+version = "0.1.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+build = "build.rs"
+
+[build-dependencies]
+rustc_version = "0.2"
+
+[features]
+default = ["std"]
+std = []
+nightly = []
+strict = []
